+++ packages/headless-media/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @maukode/headless-media
 
+## 1.0.7
+
+### Patch Changes
+
+- fixing import error and useMedia hook
+
+## 1.0.6
+
+### Patch Changes
+
+- export required new type
+
+## 1.0.5
+
+### Patch Changes
+
+- fixing recording
+
 ## 1.0.4
 
 ### Patch Changes